// frontend/src/components/chatWindow/messageComponents/Common.module.scss.d.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

declare const commonStyles: {
  'grid-container': string;
  'grid-visible': string;
  'icon': string;
  'text': string;
  'sender-info': string;
  'message': string;
  'user': string;
  'bot': string;
  'title': string;
  'loading': string;
  'dot': string;
  'property-container': string;
  'property-name': string;
  'property-value': string;
  'propGrid-container': string;
  'exportButton': string;
  'tagListContainer': string;
  'actionButton': string;
  'botLogo': string;
  'datatable-container': string;
  'infoDataTableContainer': string;
  'infoText': string;
  'feedbackDivPadding': string;
  'exportButtonWrapper': string;
  'exportBtnBackground': string;
  'columnWordBreak': string;
  'chart-container': string;
  'noDataMessageContainer': string;
  'markdown-content': string;
  'chatMessage': string;
  'moreDetails': string;
  'storageInsightsLink': string;
  'underline': string;
};
export default commonStyles;